rtc: pcf85053: add RTC driver for NXP PCF85053
This is the first known device to implement the Intel Bootable CPU I2C
RTC Common Device Specification. Other vendors and devices which
implement the spec may fall under this same driver in the future. This
initial driver only supports reading the current time.

The register layout aligns to the legacy MC146818B layout, so existing
constants are leveraged where possible. This driver is not intended for
use by the host x86 CPU, but rather for external controllers such as
the BMC which need access to the RTC over I2C.

Signed-off-by: Jonathan Doman <jonathan.doman@intel.com>