Add delay between SCI interrupt assert & de-assert
To raise the SCI interrupt BMC will assert the SCI VW wire
and de-assert it immediately.

But there seems to be some issue on host(Might be silicon bug
which BIOS team is following up) due to which SCI handler is
not getting triggered in ASL code.

This commit add 6us delay between assert SCI interrupt and
de-assert in BMC side so that ASL code can trigger the SCI
handler.

This is kind of workaround for now and will be taken out
depending on outcome of above issue from bios/silicon side.

Tested:
ASL code can trigger the SCI handler and process the data.

Signed-off-by: AppaRao Puli <apparao.puli@linux.intel.com>