mfd: syscon: Use regmap with raw_spinlock for ast2x00-scu
The irq-aspeed-scu-ic driver uses syscon regmap inside its irq_chip
implementation.
Some of the functions are called by IRQ core under raw_spinlock, since
the regmap config is controlled by the generic syscon driver, in order
to silence the warning:
[    0.687661] [ BUG: Invalid wait context ]
we need to make sure that the regmap is configured in a mode that uses
raw_spinlock.

NOTE: It is just a workaround - it will not be upstreamed, proper fix
would require either adding a new syscon DTS property, or converting the
irq-aspeed-scu-ic driver to not use generic syscon.
For more information, see the discussion related to a different driver:
Link: https://lore.kernel.org/lkml/20210825205041.927788-1-vladimir.oltean@nxp.com/

Signed-off-by: Iwona Winiarska <iwona.winiarska@intel.com>
Change-Id: Ie1e94e3799147886c8c9703672ae1657874c366a