Add Support for Circular Buffer Overflow
This commit handles the fix for circular buffer overflow that occurs
during read operation when host to bmc read pointer reaches
the end of the circular buffer. Since memcpy does not check for
overflow, an if condition is added to check if the host to bmc read
pointer has reached the end of the buffer.

TESTED:

 - SMM Driver Capsule sent successfully.
 - Complete Telemetry Logs collected.

   busctl call xyz.openbmc_project.MMBI_0.Seamless
               /xyz/openbmc_project/smm_runtime_update
               xyz.openbmc_project.smm_runtime_update
               StartFwUpdate s <capsule>

Signed-off-by: Gayathri Leburu <gayathri.leburu@intel.com>