i3c: mctp: Add IOCTL to set own EID
The driver doesn't implement MCTP logic, which means that it requires
its clients to prepare valid MCTP over I3C packets.
Because we support MCTP stacks in both kernel and userspace, and since
MCTP protocol is a part of userspace, we need userspace to update own
EID information in kernel to provide it for in-kernel i3c-mctp
clients.

To provide own EID information for in-kernel clients, add a dedicated
IOCTL.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>