Add new IOCTl command to get MMBI config
This commit adds a new IOCTL call that provides the MMBI config
information like Host Ready Bit, Host to BMC Write pointer and
BMC to Host Read Pointer.Further, In this commit we add the Host to BMC
Read pointer and BMC to Host Write Pointer.
MMBI Test Application can be used to
initiate the IOCTl command to get the MMBI configuration.

TESTED:

 - busctl call in MMBI test Application provides the MMBI config
   Info as expected.

   busctl call xyz.openbmc_project.MMBITest
   /xyz/openbmc_project/mmbi/mmbi_0_seamless
   xyz.openbmc_project.Mmbi.Test SendIOCTLCommand i 2

   Host Ready : 1
   H2B_WP  : 0xca
   B2H_RP : 0x162
   H2B_RP : 0xca
   B2H_WP : 0x162

Signed-off-by: Gayathri Leburu <gayathri.leburu@intel.com>
Signed-off-by: moni kumar <monix.kumar@intel.com>