i3c: mctp: PEC for target mode enabled
MCTP over I3C specification requires to handle PEC for both private
read and write. PEC support in hardware does not work well thus
software implementation is needed.
PEC calculation for MCTP over I3C is done with the same way it is for
SMBus. Target device address (with RnW bit) is included to the
calculation.

Signed-off-by: Zbigniew Lukwinski <zbigniew.lukwinski@linux.intel.com>