i3c: mctp: Hardcode I3C MCTP maximum IBI payload size
There is no difference which MRL for IBI payload is configured if IBI
cannot be enabled for any reason.
Furthermore, 0 as an MRL 3rd byte means that unlimited IBI payload
size is acceptable.
We can safely hardcode the value - if IBI has not been enabled -
this setting will not have any effect.

Fixes: 7c1de25c06f3 ("i3c: mctp: Add GETMWL/SETMWL and GETMRL/SETMRL support")

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>