i3c: master: dw: Add the condition for reading the response
Reading the response each time we get any interrupt might lead
to communication faults and even to kernel panics
(in particular when end_xfer_locked() called for IBI).

TRANSFER_ERR interrupt signalization can be signalized separately
from response readiness.
Thus, not to ruin communication - the response should be read when each
of the interrupts is sygnalized.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>