dts: aspeed-bmc-intel-bhs: set the MAC12_DELAY_1G value through the DTS
For BHS, we have a patch that overrides this value
ASPEED_G6_DEF_MAC12_DELAY_1G   0x0041b7d0

We pulled in a change from the Aspeed Linux fork that allows setting
this value through the DTS. This change sets the above value through the
DTS, so we can remove the patch.

Confirmed that the MAC12_DELAY_1G register value is the same as was
applied with the patch.

Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>