i3c: mctp: check priv pointer when calling write() fops
It is possible that priv pointer is NULL when calling write() fops.
On driver removal we can guarantee that the character device can't be
opened. However, the device which is already opened remains accessible.
Which means that if write() file operations was called after driver
removal - we might be in trouble due to invalid client->priv pointer
and thus should not allow such transfer happening.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>