i3c: hub: logical bus register/unregister fix
Logical buses configuration is fetched from DT in probe but logical bus
registering is done in delayed work. Thus there is period of time when
logical bus configuration is available but not yet registered. If in this
period of time hub driver unbinding happens, logical bus unregistering is
done based on information if configuration is available. It will results
with exception (NULL pointer dereference) in i3c_master_unregister()
routine.
Fix is about providing new flag which indicates if logical bus registering
really happened. Based on this flag unregistering is done.

Signed-off-by: Zbigniew Lukwinski <zbigniew.lukwinski@linux.intel.com>