i3c: master: dw: Handle repeated IBI in target mode
As per MCTP over I3C specification if Slave is unable to service
multiple IBIs in a row - it should delay IBI notifications until prior
packet is read.
At the same time we should meet 100ms timeout to repeat an IBI if there
is no read happens.
Which means that we can't delay IBIs for more than 100ms.

Due to having the same data queue for both IBI payload and private
data AST2600 must be considered as such slave. If repeated IBI happens -
the data pushed to be read by master is read by IBI payload instead which
crashes the whole communication.

To sort it out the following mechanism is introduced:
 - IBI messages are marked with Transaction ID 0 whereas private read
 messages are non-zero.
 - The 100ms timeout is started on each data push to the targets buffer.
 It is stopped when private read happens (Transaction ID !=0).
 Otherwise the following recovery procedure applied as per vendor's
 suggestion:
   - The controller is disabled;
   - CMD queue and data FIFO are reset;
   - The controller is back enabled.
To properly process controller disablement/enablement SCL should be
toggled in. To make it work the dummy scl_in signal is being set via
ASPEED AST2600 i3cglobal registers.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>