i3c: master: dw: Revise Transaction ID concept
According to the specification User-defined TID may be between
0b000-0b111.
Whereas number of empty locations in the command buffer may be up to 256.
So this is not a proper way using TID as a command counter.
The only case where TID may be used so far is differing Master Read
transaction in target mode.
Taking this into account defining 3 TID values:
 - 0b001 - Slave IBI;
 - 0b010 - Master Read;
 - 0b011 - Master Write.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>