dt-bindings: i3c: Add AST2600 i3c controller
Add a devicetree binding for the ast2600 i3c controller hardware. This
is heavily based on the designware i3c core, plus a reset facility
and two platform-specific properties:

 - sda-pullup-ohms: to specify the value of the configurable pullup
   resistors on the SDA line

 - aspeed,global-regs: to reference the (ast2600-specific) i3c global
   register block, and the device index to use within it.

NOTE: This patch was cherry-picked as-is from the patchset which was
under review:
https://lore.kernel.org/linux-i3c/cover.1676532146.git.jk@codeconstruct.com.au/
Once merged to upstream - delta must be fixed.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>
Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>