i3c: dw: Add platform operations
The dw i3c core can be integrated into various SoC devices. Platforms
that use this core may need a little configuration that is specific to
that platform.

Add a little infrastructure to allow platform-specific behaviour: a bit
of data on struct dw_i3c_master, and two hooks to the probe and init
calls to enable this.

A future change will add new platform support that uses these hooks.

NOTE: This patch was cherry-picked as-is from the patchset which was
under review:
https://lore.kernel.org/linux-i3c/cover.1676532146.git.jk@codeconstruct.com.au/
Once merged to upstream - delta must be fixed.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>
Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>