i3c: dw: Add AST2600 platform ops
Now that we have platform-specific hooks for the dw i3c driver, add
platform support for the ASPEED AST2600 SoC.

The AST2600 has a small set of "i3c global" registers, providing
platform-level i3c configuration outside of the i3c core.

For the ast2600, we need a couple of extra setup operations:

 - on probe: find the i3c global register set and parse the SDA pullup
   resistor values

 - on init: set the pullups accordingly, and set the i3c instance IDs

NOTE: This patch was cherry-picked as-is from the patchset which was
under review:
https://lore.kernel.org/linux-i3c/cover.1676532146.git.jk@codeconstruct.com.au/
Once merged to upstream - delta must be fixed.

Signed-off-by: Jeremy Kerr <jk@codeconstruct.com.au>
Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>