i3c: master: dw: enable SW DAT support
I3C driver shall support multiple I3C target devices even the number is
greater than HW DAT depth. This enabling Software (SW) DAT is required.
SW DAT mechanism still using HW DAT but all the logic are done on DAT
structures available in the I3C driver. HW DAT are updated only in case any
transaction is going to be performed or IBI is going to be enabled or
disabled. First HW DAT slot is used for communication with all I3C target
devices on the bus for which IBI was not enabled. Rest of HW DAT slots are
used for communication with I3C target devices on the bus for which IBI was
enabled. This results with having limitation for total I3C target devices
for which IBI are enabled - (HW DAT depth - 1).

Signed-off-by: Zbigniew Lukwinski <zbigniew.lukwinski@linux.intel.com>