i3c: master: dw: Add isolate_scl_sda() platform operation for AST2600
The "enable bit" of the I3C slave mode needs to wait for the bus free time
before it can be activated by the SCL signal.
To ensure the slave controller is enabled any time it is expected - we need
to use I3C global registers to isolate the SCL/SDA.
Necessary reworks to toggle_scl_in() function to use it with
isolate_scl_sda() were also done.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>