i3c: master: dw: Unify enablement/disablement procedure in AST2600
Successful completion of enablement/disablement procedure when acting
as a target can be tracked via checking DEVICE_CTRL ENABLE bit change.
Successful enablement/disablement depends on SCL availability.
We can't guarantee that each time we want to enable/disable controller the
SCL will be in high state.
Thus, we are using global registers to set SCL high for controllers
inner interpretation.

There are number of scenarios when we need to re-enable the controller:
 - On bus initialization;
 - To clean up the queues in case of timed out transfer.

Rework disable/enable functions to ensure the SCL is always in the proper
state on enablement/disablement attempt.
Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>