soc: aspeed-spilock: Update SPI Lock/unlock logic
With current logic, we are clearing Address registers
(FMCB0/B4) to unlock ROT regions. It does unlock 64K
region but locks initial 4K region i.e. 0x0 to 0xfff
which is part of ROT-P. So, to fix this issue, we are
updating logic to handle it with Address filter control
register i.e. FMCA4.

Signed-off-by: Saitwal, Meghan <meghan.saitwal@intel.com>