i3c:master:dw: Remove rsrvd bit setting
BIT(1) of DEVICE_CTRL register is said to be reserved according
to DW I3C controller specification, which means that we are not supposed
to write it.
Since writing this bit does not even change the bit itself and does not
ruin I3C communication via DW I3C controller on AST2600 A3 - the code
can be safely removed.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>