i3c: hub: Auto mode for ports added
This commit adds auto mode for hub ports declaration in the dts.
Such ports are initally declared as smbus and based on acking or
nacking certain messages they are later set to either not defined,
i3c or kept in smbus.
The system is based around matching hubs with their dts entries.
To do that a 2 bit id is used that can be set in the hub by pulling
its MSDA1 and MSCL1 pins either high or low. From that one could
achieve a "default" hub dts configuration where for example
id equal to zero would correspond with a default entry. Then all
hubs that would have their MSDA1 and MSCL1 pins pulled low
would assume this default configuration which could have all of
its ports set to auto mode. At that point there still would be
3 static configurations per i3c bus that one could use.
Such default auto configuration in the dts could look like this:

hub1@0,0 {
    tp2367-ldo = "1.8V";
    cp1-ldo = "1.8V";
    tp2367-pullup = "2k";
    tp0145-ldo = "1.8V";
    tp0145-pullup = "2k";
    cp0-ldo = "1.8V";
    id-cp1 = <0x00>;

    target-port@0 {
        mode = "auto";
        pullup = "enabled";
        backend@0,0{
            compatible = "slave-mqueue";
            target_reg = <0x10>;
        };
    };
    target-port@1 {
        mode = "auto";
        pullup = "enabled";
        backend@0,0{
            compatible = "slave-mqueue";
            target_reg = <0x10>;
        };
    };

    ...

    target-port@7 {
        mode = "auto";
        pullup = "enabled";
        backend@0,0{
            compatible = "slave-mqueue";
            target_reg = <0x10>;
        };
    };
};

The backend property is optional and will create an i2c client (with
the declared address and backend - for now only slave-mqueue is supported)
to allow the controller to act as a target in a multi-controller
environment.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>