peci-legacy: i3c: Fixing RX length based on Packet size
We are currently using fixed RX length based on command and due
to which even in case of error response code like 0x90, we
try to copy data which is not received in RX packet. So, it
gives garbage data as Output. So, we are updating RX length
based on actual RX packet length minus headers.

Signed-off-by: Saitwal, Meghan <meghan.saitwal@intel.com>