i3c: master: dw: Use IBI_ADDR_MASK
On DW I3C IP there is an option to ignore first 3 or 4 bits of device
address when ACKing/NACKing IBI.
In practice it means that if 3 bits mask is set and the device with
address 0x09 is added to the HW DAT - the IBIs from devices with addresses
0x08-0x0F are ACKed.
Use this mechanism for the cases when there are more than 8 devices on the
bus expected to allow receiving IBIs from any device requesting it.
This allows enabling IBIs for up to 7 HW DAT free slots * 0b111 ignored
bits = 49 addresses to ACK IBIs.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>