i3c: master: dw: Interfere into HW dat as late as possible
There is no need to interfere into HW DAT slots in case there is no
option to proceed I3C frame e.g. due to lack of transfers or invalid
address or any other kind of error.
Fill in the HW DAT common slot only when we are sure than we can queue
the command.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>