i3c: master: dw: utilize any available HW DAT slot
Using only slot [0] of HW DAT for all the communication with the devices
which are not in HW DAT is not the best way from the performance
perspective as more than half of the slots stay empty.

Use any available HW DAT slots to process the I3C frames as smoothly as
posslble (only the slots reserved for IBI mask remain unchanged).

To make it work the following case was considered:
 - In case all the slots in HW DAT are occupied and there is frame/command
 to be processed - put the thread in the queue until some HW DAT slot is
 released.

Signed-off-by: Oleksandr Shulzhenko <oleksandr.shulzhenko.viktorovych@intel.com>