Always disable CONFIG_DEVMEM by default
Explicitly turn off DEVMEM. It was discovered that DEVMEM was leaking
into release builds, for reasons yet to be identified. This feature is
considered a violation of Intel security practices. It is only
expected to exist in development kernels.

By explicitly disabling DEVMEM the hope is to identify the point the
feature is being enabled.

Signed-off-by: Johnathan Mantey <johnathanx.mantey@intel.com>