peci: core: remove retries on 0x82 completion codes
The CPU will return 0x82 completion code when the PECI command cannot be
completed because the CPU is asleep. Recently, the BIOS enabled Package
C6 (PC6) by default, so we have started getting a lot more 0x82
responses.

Because we retry on 0x82 for up to 700ms and many commands return 0x82,
we are hitting an issue where even PECI commands that aren't affected by
PC6 are delayed, sometimes for many seconds, because other commands
ahead of it are being retried.

This removes the retry for 0x82. Since the CPU is asleep, there is no
guarantee that it will be awake on the next retry, so in most cases,
it's causing a delay without any benefit. This removes PECI delays
caused by PC6 and gives the calling application a quicker response that
the CPU is asleep.

Signed-off-by: Jason M. Bills <jason.m.bills@intel.com>