peci-legacy: i3c: Avoid modifying in-out parameter
msg->rx_len is an in-out parameter and it was used
to capture the actual received message length on the
PECI_I3C bus. However, in case of a command retry,
the msg->rx_len gets modified. This results in
erroneous rx len in subsequent retry packets.

Fix this by capturing actual received length in a
local variable.

Signed-off-by: Sumanth Bhat <sumanth.bhat@intel.com>