soc: aspeed: espi-mmbi: Fix buffer overflow issue in MMBI driver
Added fix for possible vulnerability in MMBI driver w.r.t. buffer
overflow.

There is a possible buffer overflow when the summation of input variable
len and sizeof(hdr) variables exceeds the ULONG_MAX. All the variables
(len, sizeof and avail_buff) are of size_t which is Unsigned Long.

Signed-off-by: Kiran Kumar Subramanian <kiran.kumar.subramanian@intel.com>